+++ admin-blz/src/main/java/com/litao/query/NoticeQuery.java
@@ -19,4 +19,7 @@ public class NoticeQuery extends Query {
     @Schema(description = "通知标题")
     private String title;
 
+    @Schema(description = "通知内容")
+    private String content;
+
 }