+++ admin-core/src/main/java/com/litao/rbac/dao/SysRoleMenuDao.java
@@ -3,6 +3,9 @@ package com.litao.rbac.dao;
 import com.litao.mybatis.dao.BaseDao;
 import com.litao.rbac.entity.SysRoleMenuEntity;
 import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
 
 
 /**
@@ -12,4 +15,8 @@ import org.apache.ibatis.annotations.Mapper;
  */
 @Mapper
 public interface SysRoleMenuDao extends BaseDao<SysRoleMenuEntity> {
+    /**
+     * 根据角色ID，获取菜单ID列表
+     */
+    List<Long> getMenuIdList(@Param("roleId") Long roleId);
 }