+++ admin-core/src/main/java/com/litao/rbac/query/SysRoleQuery.java
@@ -1,0 +1,16 @@
+package com.litao.rbac.query;
+
+import com.litao.common.query.Query;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+/**
+* 角色查询
+* */
+@Data
+@EqualsAndHashCode(callSuper = false)
+@Schema(description = "角色查询")
+public class SysRoleQuery extends Query {
+    @Schema(description = "角色名称")
+    private String name;
+}