+++ admin-core/src/main/java/com/litao/rbac/query/SysRoleUserQuery.java
@@ -1,0 +1,17 @@
+package com.litao.rbac.query;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+/**
+ *用户角色查询
+ */
+
+@Data
+@EqualsAndHashCode(callSuper = false)
+@Schema(description = "分配角色查询")
+public class SysRoleUserQuery extends SysUserQuery {
+    @Schema(description = "角色ID")
+    private Long roleId;
+}