+++ admin-core/src/main/java/com/litao/rbac/service/SysUserRoleService.java
@@ -12,5 +12,43 @@ import java.util.List;
  * @author
  */
 public interface SysUserRoleService extends BaseService<SysUserRoleEntity> {
+    /**
+     * 保存或修改
+     * @param userId 用户ID
+     * @param roleIdList 角色列表
+     */
+    void saveOrUpdate(Long userId,List<Long> roleIdList);
+
+    /**
+     * 分配角色给用户列表
+     * @param roleId 角色ID
+     * @param userIDList 用户ID列表
+     */
+    void saveUserList(Long roleId,List<Long> userIDList);
+
+    /**
+     * 根据角色id列表，删除用户角色关系
+     * @param roleIdList 角色id列表
+     */
+    void deleteByRoleIdList(List<Long> roleIdList);
+
+    /**
+     * 根据用户id列表，删除用户关系角色
+     * @param userIdList 用户id列表
+     */
+    void deleteByUserIdList(List<Long> userIdList);
+
+    /**
+     * 根据角色id、用户id列表，删除用户角色关系
+     * @param roleId 角色id
+     * @param userIdList 用户id列表
+     */
+    void deleteByUserIdList(Long roleId,List<Long> userIdList);
+
+    /**
+     * 角色ID列表
+     * @param id
+     * @return
+     */
     List<Long> getRoleIdList(Long id);
 }