+++ admin-core/src/main/java/com/litao/rbac/service/impl/SysUserRoleServiceImpl.java
@@ -1,12 +1,16 @@
 package com.litao.rbac.service.impl;
 
+import cn.hutool.core.collection.CollUtil;
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.litao.mybatis.service.impl.BaseServiceImpl;
 import com.litao.rbac.dao.SysUserRoleDao;
 import com.litao.rbac.entity.SysUserRoleEntity;
 import com.litao.rbac.service.SysUserRoleService;
 import org.springframework.stereotype.Service;
 
+import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 
 
 /**
@@ -21,4 +25,59 @@ public class SysUserRoleServiceImpl extends BaseServiceImpl<SysUserRoleDao, SysU
         List<Long> roleIdList = baseMapper.getRoleIdList(id);
         return roleIdList;
     }
+
+    @Override
+    public void saveOrUpdate(Long userId, List<Long> roleIdList) {
+        // 数据库角色ID列表
+        List<Long> dbRoleIdList = getRoleIdList(userId);
+        // 需要新增的角色ID
+        Collection<Long> insertRoleIdList = CollUtil.subtract(roleIdList, dbRoleIdList);
+        if (CollUtil.isNotEmpty(insertRoleIdList)) {
+            List<SysUserRoleEntity> roleList = insertRoleIdList.stream().map(roleId -> {
+                SysUserRoleEntity entity = new SysUserRoleEntity();
+                entity.setUserId(userId);
+                entity.setRoleId(roleId);
+                return entity;
+            }).collect(Collectors.toList());
+
+            // 批量新增
+            saveBatch(roleList);
+        }
+        // 需要删除的角色ID
+        Collection<Long> deleteRoleIdList = CollUtil.subtract(dbRoleIdList, roleIdList);
+        if (CollUtil.isNotEmpty(deleteRoleIdList)) {
+            LambdaQueryWrapper<SysUserRoleEntity> queryWrapper = new LambdaQueryWrapper<>();
+            remove(queryWrapper.eq(SysUserRoleEntity::getUserId, userId).in(SysUserRoleEntity::getRoleId, deleteRoleIdList));
+        }
+    }
+
+    @Override
+    public void saveUserList(Long roleId, List<Long> userIDList) {
+        List<SysUserRoleEntity> list = userIDList.stream().map(userId -> {
+            SysUserRoleEntity entity = new SysUserRoleEntity();
+            entity.setUserId(userId);
+            entity.setRoleId(roleId);
+            return entity;
+        }).collect(Collectors.toList());
+        // 批量新增
+        saveBatch(list);
+    }
+
+    @Override
+    public void deleteByRoleIdList(List<Long> roleIdList) {
+        remove(new LambdaQueryWrapper<SysUserRoleEntity>().in(SysUserRoleEntity::getRoleId, roleIdList));
+    }
+
+    @Override
+    public void deleteByUserIdList(List<Long> userIdList) {
+        remove(new LambdaQueryWrapper<SysUserRoleEntity>().in(SysUserRoleEntity::getUserId, userIdList));
+    }
+
+    @Override
+    public void deleteByUserIdList(Long roleId, List<Long> userIdList) {
+        LambdaQueryWrapper<SysUserRoleEntity> queryWrapper = new LambdaQueryWrapper<>();
+        remove(queryWrapper.eq(SysUserRoleEntity::getRoleId, roleId).in(SysUserRoleEntity::getUserId, userIdList));
+    }
+
+
 }