+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+#  (2023-05-01)
+
+
+### Features
+
+* **custom:** 创建 admin-base 基础核心模块 669cf09
+* **custom:** 创建父工程，添加Commitizen ea53068
+* **custom:** 创建后台管理系统核心功能模块 admin-core d1bbb94
+* **custom:** 登录事件监听，保存登录日志，登录日志查询 3bbf7a8
+* **custom:** 服务启动模块完成 455b625
+* **custom:** 后台首页统计接口完成 744a860
+* **custom:** 获取登录用户信息 931e6f5
+* **custom:** 基础核心模块 common 通用类 19eaf0a
+* **custom:** 基础核心模块 security 相关配置 6c4786f
+* **custom:** 基础核心模块MybatisPlus相关配置 f3b535d
+* **custom:** 配置项目整体依赖管理 43811f2
+* **custom:** 通知功能的批量删除操作 3e3f0b8
+* **custom:** 修改用户密码功能 2fa764d
+* **custom:** 验证码登录功能 873d049
+* **custom:** 业务模块：通知模块功能开发 b24a39f
+* **custom:** 用户分页查询、添加、修改、删除、导出、导入功能实现 8793e3c
+* **custom:** 用户分页查询修改完善 d339f94
+* **custom:** 账号密码登录完成 e289bde
+* **custom:** 柱状统计图添加分类参数 2877116
+* **custom:** common 包结构搭建 092ab31
+* **custom:** rabc核心功能基础代码准备 7797300
+
+
+
 #  (2023-04-23)